Add src/clustertree.py
Contains classes ClusterTree and ClusterTreeNode classes which together contain the logic for Cluster/Instance tree family of multi label learning methods. In cluster trees, points are repeatedly clustered to give a tree. Point representations may vary (ex: label vectors or the feature vectors). Internal nodes model the conditional probability of a given sample belonging to one of the child node's clusters. This is treated as a multiclass classification problem. Leafs model the label distribution, once a sample is routed to one.

CraftML can be instantiated as a ClusterTree.