Add src/notebooks/eda_labels.ipynb Add src/notebooks/eda_tsne_visualization.ipynb
Add notebooks containing code for EDA of MLC datasets.
eda_labels contains various graphs and statistics.
eda_tsne_visualization contains interactive visualizations of labels in feature space, labels in labelset space and labelsets in feature space. Purpose is to determine if labels/labelsets are localized in X space, and if labels are localized in Y space.